rename cpp-dump.hpp -> cpp_dump.hpp
