+++ dump.hpp
@@ -1,5 +1,5 @@
 #pragma once
 #pragma message(                                                                                   \
-    "WARNING: The 'dump.hpp' header is deprecated. Include the 'cpp-dump.hpp' header instead."     \
+    "WARNING: The 'dump.hpp' header is deprecated. Include the 'cpp_dump.hpp' header instead."     \
 )
-#include "./cpp-dump.hpp"
+#include "./cpp_dump.hpp"
