+++ readme/customizable-colors.cpp
@@ -12,7 +12,7 @@
 #include <variant>
 #include <vector>
 
-#include "../cpp-dump.hpp"
+#include "../cpp_dump.hpp"
 namespace cp = cpp_dump;
 
 int main() {
