+++ readme/customize-dump.cpp
@@ -3,7 +3,7 @@
 #include <set>
 #include <vector>
 
-#include "../cpp-dump.hpp"
+#include "../cpp_dump.hpp"
 namespace cp = cpp_dump;
 
 void my_func() {
