+++ readme/formatting-with-manipulators.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <vector>
 
-#include "../cpp-dump.hpp"
+#include "../cpp_dump.hpp"
 namespace cp = cpp_dump;
 
 int main() {
