+++ readme/introduction.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <vector>
 
-#include "../cpp-dump.hpp"
+#include "../cpp_dump.hpp"
 
 int main() {
   std::clog << std::endl;
