+++ readme/user-defined-enum.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "../cpp-dump.hpp"
+#include "../cpp_dump.hpp"
 
 enum class enum_A { a, b, c };
 CPP_DUMP_DEFINE_EXPORT_ENUM(enum_A, enum_A::a, enum_A::b, enum_A::c);
