+++ test/color_non_variable_test.cpp
@@ -1,4 +1,4 @@
-#include "../cpp-dump.hpp"
+#include "../cpp_dump.hpp"
 namespace cp = cpp_dump;
 
 #include <array>
