+++ test/odr_test.cpp
@@ -1,5 +1,5 @@
 #include "./odr_test.hpp"
 
-#include "../cpp-dump.hpp"
+#include "../cpp_dump.hpp"
 
 void odr_test() { cpp_dump(class_a(), class_b(), enum_a::s); }
